modifications mineures

+++ code/Hadrien/MDMA/src/objects/midimanager.cpp
@@ -48,7 +48,7 @@ void MidiManager::createPort(QString name)
 
 	// We need the name of the MIDI port in LPCWSTR, this this awfull part of code
 	int len;
-	int slength = (int)name.length() + 1;
+	int slength = name.length() + 1;
 	len = MultiByteToWideChar(CP_ACP, 0, name.toStdString(), slength, 0, 0);
 	wchar_t* buf = new wchar_t[len];
 	MultiByteToWideChar(CP_ACP, 0, name.c_str(), slength, buf, len);
